Add `fancyvrb` aspect to Mathematica language
Before this commit, when first used cell was a `formatted` one, an error
was printed, informing about undefined key: `morefvcmdparams`.

It seems listings `fancyvrb` aspect is loaded automatically only when
`fancyvrb` key is used, but not when other keys defined in this aspect
are used.

Fixed by explicitly requiring `fancyvrb` aspect in `base` `Mathematica`
language definition.